Implement user authentication and authorization; add registration and login functionality; enhance comment and post management with role-based access control

+++ src/main/java/com/mountblue/blog/blog_application/service/CommentService.java
@@ -1,14 +1,13 @@
 package com.mountblue.blog.blog_application.service;
 
 import com.mountblue.blog.blog_application.dto.CommentDto;
-import com.mountblue.blog.blog_application.dto.PostDto;
 import com.mountblue.blog.blog_application.exception.NoDataFound;
 import com.mountblue.blog.blog_application.model.Comment;
 import com.mountblue.blog.blog_application.model.Post;
-import com.mountblue.blog.blog_application.model.User;
 import com.mountblue.blog.blog_application.repository.CommentRepository;
 import com.mountblue.blog.blog_application.repository.PostRepository;
-import com.mountblue.blog.blog_application.repository.UserRepository;
+import com.mountblue.blog.blog_application.util.Utils;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -16,14 +15,14 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
+    private final Utils utils;
 
     public CommentService(CommentRepository commentRepository,
                           PostRepository postRepository,
-                          UserRepository userRepository){
+                          Utils utils){
         this.commentRepository = commentRepository;
         this.postRepository = postRepository;
-        this.userRepository = userRepository;
+        this.utils = utils;
     }
 
     public void addComment(CommentDto commentDto){
@@ -42,35 +41,59 @@ public class CommentService {
         }
     }
 
-    public Long deleteComment(long commentId) {
+    public Long deleteComment(long commentId, String userEmail, boolean isAdmin) {
         try {
             Comment comment = commentRepository.getReferenceById(commentId);
+            Post post = comment.getPost();
+
+            if(!isAdmin && utils.isAuthor(userEmail,post)){
+                throw new AccessDeniedException("You are not authorized to delete this comment.");
+            }
+
             Long postId = comment.getPost().getId();
             commentRepository.delete(comment);
+
             return postId;
-        } catch (Exception e) {
+        } catch (AccessDeniedException e) {
+            throw e;
+        }
+        catch (Exception e) {
             throw new RuntimeException("Failed to delete comment. Please try again later. Error: " + e.getMessage());
         }
     }
 
-    public Comment fetchCommentById(Long commentId) {
-        return commentRepository.findById(commentId).orElseThrow(
+    public Comment fetchCommentById(Long commentId, String userEmail, boolean isAdmin) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow(
                 ()-> new NoDataFound("Comment Not Found with id " + commentId)
         );
+
+        if(!isAdmin && !utils.isAuthor(userEmail,comment.getPost())){
+            throw new AccessDeniedException("You are not authorized to edit this comment.");
+        }
+
+        return comment;
     }
 
 
-    public Long updateComment(CommentDto commentDto) {
+    public Long updateComment(CommentDto commentDto, String userEmail, boolean isAdmin) {
         try {
             Comment comment = commentRepository.findById(commentDto.getId()).orElseThrow(
                     () -> new NoDataFound("Comment Not Found with id " + commentDto.getId())
             );
-            Long postId = comment.getPost().getId();
+            Post post = comment.getPost();
+
+            if(!isAdmin && !utils.isAuthor(userEmail,post)){
+                throw new AccessDeniedException("You are not authorized to update this comment.");
+            }
+
             comment.setComment(commentDto.getComment());
             commentRepository.save(comment);
 
-            return postId;
-        } catch (Exception e) {
+            return post.getId();
+        } catch (AccessDeniedException e) {
+            throw e;
+        }
+        catch (Exception e) {
             throw new RuntimeException("Failed to update comment. Please try again later. Error: " + e.getMessage());
         }
     }
